# security/openapi.json
{"openapi":"3.1.0","info":{"title":"System Zarządzania Kliniką Weterynaryjną","description":"Aplikacja wykorzystująca FastAPI, SQLAlchemy oraz defensywne programowanie.","version":"1.0.0"},"paths":{"/users/register":{"post":{"tags":["users"],"summary":"Register Client","operationId":"register_client_users_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/create-doctor":{"post":{"tags":["users"],"summary":"Create Doctor","operationId":"create_doctor_users_create_doctor_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DoctorCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/create-consultant":{"post":{"tags":["users"],"summary":"Create Consultant","operationId":"create_consultant_users_create_consultant_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultantCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/":{"get":{"tags":["users"],"summary":"Read Users","operationId":"read_users_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserOut"},"type":"array","title":"Response Read Users Users  Get"}}}}}}},"/users/login":{"post":{"tags":["users"],"summary":"Login","description":"Endpoint logowania z blokadą po 5 nieudanych próbach (15 min).","operationId":"login_users_login_post","parameters":[{"name":"force_provision","in":"query","required":false,"schema":{"type":"boolean","description":"Jeśli True, wymusza ponowną konfigurację TOTP","default":false,"title":"Force Provision"},"description":"Jeśli True, wymusza ponowną konfigurację TOTP"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/setup-totp":{"post":{"tags":["users"],"summary":"Setup Totp","operationId":"setup_totp_users_setup_totp_post","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string","title":"Email"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/confirm-totp":{"post":{"tags":["users"],"summary":"Confirm Totp","description":"Endpoint potwierdzający konfigurację TOTP.\n\nPrzyjmuje:\n  - email: identyfikator użytkownika,\n  - totp_code: 6-cyfrowy kod generowany przez aplikację TOTP.\n\nJeśli użytkownik został znaleziony i posiada skonfigurowany totp_secret,\na przesłany kod jest poprawny, ustawiamy flagę totp_confirmed na True i zwracamy komunikat w polu detail.","operationId":"confirm_totp_users_confirm_totp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmTOTP"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/appointments/":{"post":{"tags":["appointments"],"summary":"Create Appointment","operationId":"create_appointment_appointments__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Appointment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["appointments"],"summary":"Read Appointments","operationId":"read_appointments_appointments__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Appointment"},"title":"Response Read Appointments Appointments  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/appointments/{appointment_id}":{"get":{"tags":["appointments"],"summary":"Read Appointment","operationId":"read_appointment_appointments__appointment_id__get","parameters":[{"name":"appointment_id","in":"path","required":true,"schema":{"type":"integer","title":"Appointment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Appointment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["appointments"],"summary":"Update Appointment","operationId":"update_appointment_appointments__appointment_id__put","parameters":[{"name":"appointment_id","in":"path","required":true,"schema":{"type":"integer","title":"Appointment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Appointment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["appointments"],"summary":"Delete Appointment","operationId":"delete_appointment_appointments__appointment_id__delete","parameters":[{"name":"appointment_id","in":"path","required":true,"schema":{"type":"integer","title":"Appointment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Appointment"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/animals/":{"post":{"tags":["animals"],"summary":"Create Animal","operationId":"create_animal_animals__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnimalCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Animal"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["animals"],"summary":"Read Animals","operationId":"read_animals_animals__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Animal"},"title":"Response Read Animals Animals  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/animals/{animal_id}":{"get":{"tags":["animals"],"summary":"Read Animal","operationId":"read_animal_animals__animal_id__get","parameters":[{"name":"animal_id","in":"path","required":true,"schema":{"type":"integer","title":"Animal Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Animal"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["animals"],"summary":"Update Animal","operationId":"update_animal_animals__animal_id__put","parameters":[{"name":"animal_id","in":"path","required":true,"schema":{"type":"integer","title":"Animal Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnimalUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Animal"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["animals"],"summary":"Delete Animal","operationId":"delete_animal_animals__animal_id__delete","parameters":[{"name":"animal_id","in":"path","required":true,"schema":{"type":"integer","title":"Animal Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Animal"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/medical_records/":{"post":{"tags":["medical_records"],"summary":"Create Record","operationId":"create_record_medical_records__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MedicalRecordCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MedicalRecord"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["medical_records"],"summary":"List Records","operationId":"list_records_medical_records__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MedicalRecord"},"title":"Response List Records Medical Records  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/medical_records/{record_id}":{"get":{"tags":["medical_records"],"summary":"Get Record","operationId":"get_record_medical_records__record_id__get","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","title":"Record Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MedicalRecord"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["medical_records"],"summary":"Put Record","operationId":"put_record_medical_records__record_id__put","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","title":"Record Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MedicalRecordUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MedicalRecord"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["medical_records"],"summary":"Del Record","operationId":"del_record_medical_records__record_id__delete","parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","title":"Record Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/invoices/":{"get":{"tags":["invoices"],"summary":"Read Invoices","operationId":"read_invoices_invoices__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/weight-logs/":{"get":{"tags":["weight-logs"],"summary":"Read Logs","operationId":"read_logs_weight_logs__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WeightLogOut"},"title":"Response Read Logs Weight Logs  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["weight-logs"],"summary":"Add Log","operationId":"add_log_weight_logs__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeightLogCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeightLogOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/weight-logs/{log_id}":{"get":{"tags":["weight-logs"],"summary":"Read Log","operationId":"read_log_weight_logs__log_id__get","parameters":[{"name":"log_id","in":"path","required":true,"schema":{"type":"integer","title":"Log Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeightLogOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["weight-logs"],"summary":"Remove Log","operationId":"remove_log_weight_logs__log_id__delete","parameters":[{"name":"log_id","in":"path","required":true,"schema":{"type":"integer","title":"Log Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeightLogOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Animal":{"properties":{"name":{"type":"string","title":"Name","description":"Imię zwierzęcia"},"species":{"type":"string","title":"Species","description":"Gatunek, np. pies, kot"},"breed":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Breed","description":"Rasa zwierzęcia"},"gender":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gender","description":"Płeć zwierzęcia, np. male lub female"},"birth_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Date","description":"Data urodzenia zwierzęcia"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age","description":"Wiek zwierzęcia (opcjonalnie)"},"weight":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Weight","description":"Waga zwierzęcia w kg"},"microchip_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Microchip Number","description":"Numer mikroczipa"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes","description":"Dodatkowe uwagi"},"owner_id":{"type":"integer","title":"Owner Id","description":"ID właściciela zwierzęcia"},"last_visit":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Visit","description":"Data ostatniej wizyty u weterynarza"},"id":{"type":"integer","title":"Id","description":"Unikalny identyfikator"},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Data utworzenia rekordu"},"updated_at":{"type":"string","format":"date-time","title":"Updated At","description":"Data ostatniej modyfikacji"}},"type":"object","required":["name","species","owner_id","id","created_at","updated_at"],"title":"Animal"},"AnimalCreate":{"properties":{"name":{"type":"string","title":"Name","description":"Imię zwierzęcia"},"species":{"type":"string","title":"Species","description":"Gatunek, np. pies, kot"},"breed":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Breed","description":"Rasa zwierzęcia"},"gender":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gender","description":"Płeć zwierzęcia, np. male lub female"},"birth_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Date","description":"Data urodzenia zwierzęcia"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age","description":"Wiek zwierzęcia (opcjonalnie)"},"weight":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Weight","description":"Waga zwierzęcia w kg"},"microchip_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Microchip Number","description":"Numer mikroczipa"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes","description":"Dodatkowe uwagi"},"owner_id":{"type":"integer","title":"Owner Id","description":"ID właściciela zwierzęcia"},"last_visit":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Visit","description":"Data ostatniej wizyty u weterynarza"}},"type":"object","required":["name","species","owner_id"],"title":"AnimalCreate"},"AnimalUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name","description":"Imię zwierzęcia"},"species":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Species","description":"Gatunek"},"breed":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Breed","description":"Rasa"},"gender":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gender","description":"Płeć"},"birth_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Date","description":"Data urodzenia"},"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age","description":"Wiek"},"weight":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Weight","description":"Waga w kg"},"microchip_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Microchip Number","description":"Numer mikroczipa"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes","description":"Uwagi"},"owner_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Owner Id","description":"ID właściciela"},"last_visit":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Visit","description":"Data ostatniej wizyty"}},"type":"object","title":"AnimalUpdate"},"Appointment":{"properties":{"visit_datetime":{"type":"string","format":"date-time","title":"Visit Datetime","description":"Data i czas wizyty"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason","description":"Powód wizyty lub rodzaj usługi"},"treatment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Treatment","description":"Zastosowane leczenie podczas wizyty"},"priority":{"type":"string","title":"Priority","description":"Priorytet wizyty: normalna, pilna, nagła","default":"normalna"},"weight":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Weight","description":"Waga zwierzęcia w dniu wizyty (kg)"},"doctor_id":{"type":"integer","title":"Doctor Id","description":"ID lekarza obsługującego wizytę"},"animal_id":{"type":"integer","title":"Animal Id","description":"ID zwierzęcia będącego pacjentem"},"owner_id":{"type":"integer","title":"Owner Id","description":"ID właściciela zwierzęcia"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes","description":"Dodatkowe uwagi"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["visit_datetime","doctor_id","animal_id","owner_id","id","created_at","updated_at"],"title":"Appointment"},"AppointmentCreate":{"properties":{"visit_datetime":{"type":"string","format":"date-time","title":"Visit Datetime","description":"Data i czas wizyty"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason","description":"Powód wizyty lub rodzaj usługi"},"treatment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Treatment","description":"Zastosowane leczenie podczas wizyty"},"priority":{"type":"string","title":"Priority","description":"Priorytet wizyty: normalna, pilna, nagła","default":"normalna"},"weight":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Weight","description":"Waga zwierzęcia w dniu wizyty (kg)"},"doctor_id":{"type":"integer","title":"Doctor Id","description":"ID lekarza obsługującego wizytę"},"animal_id":{"type":"integer","title":"Animal Id","description":"ID zwierzęcia będącego pacjentem"},"owner_id":{"type":"integer","title":"Owner Id","description":"ID właściciela zwierzęcia"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes","description":"Dodatkowe uwagi"}},"type":"object","required":["visit_datetime","doctor_id","animal_id","owner_id"],"title":"AppointmentCreate","description":"Do tworzenia wizyty – wszystkie pola z AppointmentBase są wymagane/ustawiane."},"AppointmentUpdate":{"properties":{"visit_datetime":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Visit Datetime","description":"Data i czas wizyty"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason","description":"Powód wizyty lub rodzaj usługi"},"treatment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Treatment","description":"Zastosowane leczenie podczas wizyty"},"priority":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Priority","description":"Priorytet wizyty"},"weight":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Weight","description":"Waga zwierzęcia w dniu wizyty (kg)"},"doctor_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Doctor Id","description":"ID lekarza obsługującego wizytę"},"animal_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Animal Id","description":"ID zwierzęcia będącego pacjentem"},"owner_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Owner Id","description":"ID właściciela zwierzęcia"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes","description":"Dodatkowe uwagi"}},"type":"object","title":"AppointmentUpdate","description":"Do częściowej aktualizacji wizyty – wszystkie pola opcjonalne."},"ClientCreate":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"type":"string","title":"Role"},"phone_number":{"type":"string","title":"Phone Number"},"address":{"type":"string","title":"Address"},"postal_code":{"type":"string","title":"Postal Code"}},"type":"object","required":["first_name","last_name","email","password","role","phone_number","address","postal_code"],"title":"ClientCreate"},"ConfirmTOTP":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"totp_code":{"type":"string","title":"Totp Code"}},"type":"object","required":["email","totp_code"],"title":"ConfirmTOTP"},"ConsultantCreate":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"type":"string","title":"Role"}},"type":"object","required":["first_name","last_name","email","password","role"],"title":"ConsultantCreate"},"DoctorCreate":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"type":"string","title":"Role"},"specialization":{"type":"string","title":"Specialization"},"permit_number":{"type":"string","title":"Permit Number"}},"type":"object","required":["first_name","last_name","email","password","role","specialization","permit_number"],"title":"DoctorCreate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MedicalRecord":{"properties":{"description":{"type":"string","title":"Description"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"appointment_id":{"type":"integer","title":"Appointment Id"},"animal_id":{"type":"integer","title":"Animal Id"}},"type":"object","required":["description","id","created_at","appointment_id","animal_id"],"title":"MedicalRecord"},"MedicalRecordCreate":{"properties":{"description":{"type":"string","title":"Description"},"appointment_id":{"type":"integer","title":"Appointment Id"},"animal_id":{"type":"integer","title":"Animal Id"}},"type":"object","required":["description","appointment_id","animal_id"],"title":"MedicalRecordCreate"},"MedicalRecordUpdate":{"properties":{"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"appointment_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Appointment Id"},"animal_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Animal Id"}},"type":"object","title":"MedicalRecordUpdate"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"totp_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Totp Code"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserOut":{"properties":{"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"id":{"type":"integer","title":"Id"},"role":{"type":"string","title":"Role"}},"type":"object","required":["first_name","last_name","email","id","role"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WeightLogCreate":{"properties":{"animal_id":{"type":"integer","title":"Animal Id","description":"ID zwierzęcia"},"weight":{"type":"number","exclusiveMinimum":0.0,"title":"Weight","description":"Waga zwierzęcia w kg"},"recorded_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Recorded At"}},"type":"object","required":["animal_id","weight"],"title":"WeightLogCreate"},"WeightLogOut":{"properties":{"animal_id":{"type":"integer","title":"Animal Id","description":"ID zwierzęcia"},"weight":{"type":"number","exclusiveMinimum":0.0,"title":"Weight","description":"Waga zwierzęcia w kg"},"recorded_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Recorded At","description":"Data i czas pomiaru; domyślnie ustawiana przez bazę"},"id":{"type":"integer","title":"Id","description":"ID rekordu"}},"type":"object","required":["animal_id","weight","id"],"title":"WeightLogOut"}}}}